fix duplicate RoomSection
